backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.9.0
python-multipart==0.0.9

# AI Services
google-generativeai==0.8.0

# Database
supabase==2.10.0
